refactor(server): isolate convert pandas implem
