+++ server/tests/steps/test_convert.py
@@ -6,7 +6,8 @@ import pytest
 from pandas import DataFrame, Timestamp
 
 from tests.utils import assert_dataframes_equals
-from weaverbird.steps import ConvertStep
+from weaverbird.backends.pandas_executor.steps.convert import execute_convert
+from weaverbird.pipeline.steps import ConvertStep
 
 
 @pytest.fixture
@@ -15,23 +16,24 @@ def sample_df() -> DataFrame:
 
 
 def test_convert_to_float(sample_df: DataFrame):
-    df_result = ConvertStep(name='convert', columns=['value'], data_type='float').execute(sample_df)
+    step = ConvertStep(name='convert', columns=['value'], data_type='float')
+    df_result = execute_convert(step, sample_df)
 
     expected_result = DataFrame({'value': [41.0, 42.0, 43.5, 43.6, None, None]})
     assert_dataframes_equals(df_result, expected_result)
 
 
 def test_convert_to_integer(sample_df: DataFrame):
-    df_result = ConvertStep(name='convert', columns=['value'], data_type='integer').execute(
-        sample_df
-    )
+    step = ConvertStep(name='convert', columns=['value'], data_type='integer')
+    df_result = execute_convert(step, sample_df)
 
     expected_result = DataFrame({'value': [41, 42, 43, 43, None, None]})
     assert_dataframes_equals(df_result, expected_result)
 
 
 def test_convert_to_text(sample_df: DataFrame):
-    df_result = ConvertStep(name='convert', columns=['value'], data_type='text').execute(sample_df)
+    step = ConvertStep(name='convert', columns=['value'], data_type='text')
+    df_result = execute_convert(step, sample_df)
 
     expected_result = DataFrame({'value': ['41', '42', '43.5', '43.6', 'None', 'meh']})
     assert_dataframes_equals(df_result, expected_result)
@@ -50,7 +52,8 @@ def test_convert_to_datetime():
             ]
         }
     )
-    df_result = ConvertStep(name='convert', columns=['value'], data_type='date').execute(input_df)
+    step = ConvertStep(name='convert', columns=['value'], data_type='date')
+    df_result = execute_convert(step, input_df)
 
     expected_result = DataFrame(
         {
@@ -69,9 +72,8 @@ def test_convert_to_datetime():
 
 def test_convert_to_bool():
     input_df = DataFrame({'value': ['plop', 0, 0.0, 1, '', None]})
-    df_result = ConvertStep(name='convert', columns=['value'], data_type='boolean').execute(
-        input_df
-    )
+    step = ConvertStep(name='convert', columns=['value'], data_type='boolean')
+    df_result = execute_convert(step, input_df)
 
     expected_result = DataFrame({'value': [True, False, False, True, False, False]})
     assert_dataframes_equals(df_result, expected_result)
@@ -88,4 +90,4 @@ def test_benchmark_convert(benchmark):
     )
     step = ConvertStep(name='convert', columns=['date'], data_type='date')
 
-    benchmark(step.execute, df)
+    benchmark(execute_convert, step, df)
