+++ server/weaverbird/backends/pandas_executor/steps/concatenate.py
@@ -1,4 +1,5 @@
 from pandas import DataFrame
+
 from weaverbird.backends.pandas_executor.types import DomainRetriever, PipelineExecutor
 from weaverbird.pipeline.steps import ConcatenateStep
